# Conflicts:
#	Modules/Physics/Sources/Physics/PhysicsSystem.h


Former-commit-id: 85608044625c3d0a7e1d60a4b2e49821aa1da265

+++ Sources/Internal/Base/Private/Type_impl.h
@@ -108,7 +108,7 @@ private:
     Vector<Type**> types;
 };
 
-inline size_t Type::GetSize() const
+inline uint32_t Type::GetSize() const
 {
     return size;
 }
@@ -217,7 +217,7 @@ Type* Type::Init()
     static const bool needPointer = (!std::is_pointer<T>::value);
     static const bool needSeed = (std::is_base_of<Type::Seed, T>::value || std::is_same<Type::Seed, T>::value);
 
-    type.size = TypeDetail::TypeSize<T>::size;
+    type.size = static_cast<uint32_t>(TypeDetail::TypeSize<T>::size);
     type.name = typeid(T).name();
     type.stdTypeInfo = &typeid(T);
